fix variable userId int to long

+++ app/src/main/java/com/example/timerstudy/data/remote/ApiService.java
@@ -165,7 +165,7 @@ public interface ApiService {
 
     class UserInfo {
         @SerializedName("user_id")
-        public int userId;
+        public long userId;
         @SerializedName("email")
         public String email;
         @SerializedName("display_name")
@@ -201,7 +201,7 @@ public interface ApiService {
         @SerializedName("rank")
         public int rank;
         @SerializedName("user_id")
-        public int userId;
+        public long userId;
         @SerializedName("display_name")
         public String displayName;
         @SerializedName("profile_picture_url")
@@ -231,7 +231,7 @@ public interface ApiService {
         @SerializedName("task_id")
         public int taskId;
         @SerializedName("user_id")
-        public int userId;
+        public long userId;
         @SerializedName("title")
         public String title;
         @SerializedName("description")
